starts refactor of benchmark games and base Game now stores and loads values properly.
