+++ tests/tests_games/test_games_synthetic.py
@@ -1,5 +1,8 @@
 """This test module tests the SOUM Game class"""
 
+import pytest
+
+from shapiq import DummyGame
 from shapiq.games.soum import SOUM, UnanimityGame
 import numpy as np
 
@@ -76,3 +79,71 @@ def test_soum_interations():
             )
         assert soum_4.max_interaction_size == max_interaction_size
         assert soum_4.min_interaction_size == min_interaction_size
+
+
+@pytest.mark.parametrize(
+    "n, interaction, expected",
+    [
+        (
+            3,
+            (1, 2),
+            {
+                (): 0,
+                (0,): 1 / 3,
+                (1,): 1 / 3,
+                (2,): 1 / 3,
+                (0, 1): 2 / 3,
+                (0, 2): 2 / 3,
+                (1, 2): 2 / 3 + 1,
+                (0, 1, 2): 3 / 3 + 1,
+            },
+        ),
+        (
+            4,
+            (1, 2),
+            {
+                (): 0,
+                (0,): 1 / 4,
+                (1,): 1 / 4,
+                (2,): 1 / 4,
+                (3,): 1 / 4,
+                (0, 1): 2 / 4,
+                (1, 2): 2 / 4 + 1,
+                (2, 3): 2 / 4,
+                (0, 1, 2): 3 / 4 + 1,
+                (1, 2, 3): 3 / 4 + 1,
+                (0, 1, 2, 3): 4 / 4 + 1,
+            },
+        ),
+    ],
+)
+def test_dummy_game(n, interaction, expected):
+    """Test the DummyGame class."""
+    game = DummyGame(n=n, interaction=interaction)
+    for coalition in expected.keys():
+        x_input = np.zeros(shape=(n,), dtype=bool)
+        x_input[list(coalition)] = True
+        assert game(x_input)[0] == expected[coalition]
+
+    string_game = str(game)
+    assert repr(game) == string_game
+
+
+def test_dummy_game_access_counts():
+    """Test how often the game was called."""
+    game = DummyGame(n=10, interaction=(1, 2))
+    assert game.access_counter == 0
+    game(np.asarray([True, False, False, False, False, False, False, False, False, False]))
+    assert game.access_counter == 1
+    game(np.asarray([True, False, False, False, False, False, False, False, False, False]))
+    assert game.access_counter == 2
+    game(
+        np.asarray(
+            [
+                [True, False, False, False, False, False, False, False, False, False],
+                [False, True, False, False, False, False, False, False, False, False],
+                [False, False, True, False, False, False, False, False, False, False],
+            ]
+        )
+    )
+    assert game.access_counter == 5
